chore: move json config to models
