+++ alistral_cli/src/tools/radio/yumako/mod.rs
@@ -8,7 +8,6 @@ use chrono::Utc;
 use clap::Parser;
 use clap::ValueEnum;
 use interzic::models::playlist_stub::PlaylistStub;
-use interzic::models::services::listenbrainz::Listenbrainz;
 use itertools::Itertools;
 use tracing::debug;
 use tracing::error;
@@ -16,7 +15,7 @@ use tuillez::fatal_error::FatalError;
 use tuillez::fatal_error::IntoFatal;
 use tuillez::formatter::FormatWithAsync;
 use yumako_jams::RadioStream;
-use yumako_jams::json::radio::Radio;
+use yumako_jams::models::radio_file::radio::Radio;
 use yumako_jams::radio_item::RadioItem;
 use yumako_jams::radio_stream::RadioStreamaExt;
 use yumako_jams::radio_variables::RadioVariables;
