+++ yumako_jams/Cargo.toml
@@ -23,7 +23,6 @@ tracing = "0.1.41"
 extend = "1.2.0"
 tokio = { version = "1.44.0", features = ["full"] }
 rand = "0.9.0"
-sqlx = "0.8.3"
 itertools = "0.14.0"
 priority-queue = "2.3.0"
 toml = "0.8.20"
