+++ yumako_jams/src/modules/filters/booleans.rs
@@ -10,7 +10,7 @@ use tracing::debug;
 
 use crate::RadioStream;
 use crate::client::YumakoClient;
-use crate::json::radio::Radio;
+use crate::models::radio_file::radio::Radio;
 use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
 use crate::radio_stream::RadioStreamaExt;
